+++ R/firstlib.R
@@ -1,6 +1,6 @@
 # This file is a part of rOpenGov (http://ropengov.org)
 
-# Copyright (C) 2010-2021 Juuso Parkkinen, Joona Lehtomaki, Pyry Kantanen, Leo Lahti <ropengov.org>. 
+# Copyright (C) 2010-2022 Juuso Parkkinen, Joona Lehtomaki, Pyry Kantanen, Leo Lahti <ropengov.org>. 
 # All rights reserved.
 
 # This program is open source software; you can redistribute it and/or modify 
@@ -19,6 +19,6 @@
   # Sys.setlocale(locale = "UTF-8")
   # Sys.setlocale(locale = "WINDOWS-1252")
 
-  packageStartupMessage("\nhelsinki R package: tools for open data from Helsinki, Finland capital region.\nThis R package is part of rOpenGov <ropengov.org>.\nCopyright (C) 2010-2021 Juuso Parkkinen, Joona Lehtomaki, Pyry Kantanen, Leo Lahti.\nThis is free software. You are free to use, modify and redistribute it under the FreeBSD license.")
+  packageStartupMessage("\nhelsinki R package: tools for open data from Helsinki, Finland capital region.\nThis R package is part of rOpenGov <ropengov.org>.\nCopyright (C) 2010-2022 Juuso Parkkinen, Joona Lehtomaki, Pyry Kantanen, Leo Lahti.\nThis is free software. You are free to use, modify and redistribute it under the FreeBSD license.")
 
 }