+++ R/wfs_api.R
@@ -8,10 +8,6 @@
 #'    
 #'    This is a low-level function intended to be used by other higher level
 #'    functions in the package.
-#'    
-#'    Note that GET requests are used using `httpcache` meaning that requests
-#'    are cached. If you want clear cache, use [httpcache::clearCache()]. To 
-#'    turn the cache off completely, use [httpcache::cacheOff()]
 #' 
 #' @source Gracefully failing HTTP request code (slightly adapted by Pyry 
 #'    Kantanen) from RStudio community member kvasilopoulos. Many thanks!
@@ -21,6 +17,7 @@
 #'
 #' @param base.url WFS url, for example "https://kartta.hsy.fi/geoserver/wfs"
 #' @param queries List of query parameters
+#' @param timeout.s timeout in seconds, passed onto [gracefully_fail()]
 #'
 #' @return wfs_api (S3) object with the following attributes:
 #'        \describe{
@@ -34,26 +31,42 @@
 #'
 #' @examples
 #'   wfs_api(base.url = "https://kartta.hsy.fi/geoserver/wfs", 
-#'           queries = append(list("service" = "WFS", "version" = "1.0.0"), 
-#'                 list(request = "getFeature", 
-#'                      layer = "tilastointialueet:kunta4500k_2017")))
+#'           queries = c("service" = "WFS", 
+#'           "version" = "1.0.0", 
+#'           "request" = "getFeature", 
+#'           "typeName" = "ilmanlaatu:Ilmanlaatu_nyt")
+#'           )
 #'                      
 #' @importFrom xml2 read_xml xml_text xml_has_attr xml_children
-#' @importFrom httpcache GET
-#' @importFrom httr timeout user_agent modify_url http_error message_for_status
+#' @importFrom httr GET timeout user_agent modify_url http_error message_for_status
 #' @importFrom curl has_internet
 #'                      
 #' @export
-wfs_api <- function(base.url = NULL, queries) {
-  
+wfs_api <- function(base.url = NULL, queries, timeout.s = NULL) {
+
   if (is.null(base.url)) {
     stop("base.url = NULL. Please input a valid WFS url")
   }
-  
+
   if (!grepl("^http", base.url)) {
     stop("Invalid base URL")
   }
-  
+
+  if (is.null(names(queries))) {
+    message("Please input the list in name-value pairs: c(\"name1\" = \"value1\", \"name2\" = \"value2\", ...)")
+    invisible(return(NULL))
+  }
+
+  # Test if queries is a named "Named num" or "Named chr" vector
+  if (is.vector(queries, mode = "character") || is.vector(queries, mode = "numeric")) {
+    queries <- as.list(queries)
+  }
+
+  if (!is.list(queries)) {
+    message("Please input the list in name-value pairs: c(\"name1\" = \"value1\", \"name2\" = \"value2\", ...)")
+    invisible(return(NULL))
+  }
+
   # Set the user agent
   ua <- httr::user_agent("https://github.com/rOpenGov/helsinki")
   
@@ -63,39 +76,15 @@ wfs_api <- function(base.url = NULL, queries) {
   # Print out the URL
   message("Requesting response from: ", url)
   
-  # Following code snippet from RStudio Community contributor kvasilopoulos
-  # See function help for more details ________________________________________
-  try_GET <- function(x, ...) {
-    tryCatch(
-      GET(url = x, httr::timeout(20), ...),
-      error = function(e) conditionMessage(e),
-      warning = function(w) conditionMessage(w)
-    )
-  }
-  is_response <- function(x) {
-    class(x) == "response"
-  }
+  graceful_result <- gracefully_fail(url, timeout.s = NULL)
   
-  # First check internet connection
-  if (!curl::has_internet()) {
-    message("No internet connection.")
+  if (is.null(graceful_result)) {
+    message("Please check your settings or function parameters \n")
     return(invisible(NULL))
   }
-  # Then try for timeout problems
-  resp <- try_GET(url)
-  if (!is_response(resp)) {
-    message(resp)
-    return(invisible(NULL))
-  }
-  # Then stop if status > 400
-  if (httr::http_error(resp)) { 
-    httr::message_for_status(resp)
-    return(invisible(NULL))
-  }
-  # Many thanks! ______________________________________________________________
   
   # Get the response and check the response.
-  resp <- httpcache::GET(url, ua)
+  resp <- httr::GET(url, ua)
   
   # Parse the response XML content
   content <- xml2::read_xml(resp$content)
@@ -109,10 +98,10 @@ wfs_api <- function(base.url = NULL, queries) {
   
   # Even if GET and read_xml work properly, content can be unusable
   if (any(truth_table)) {
-    warning(paste("Please check your download parameters.", 
+    message(paste("Please check your download parameters.", 
                   "Detected following exception(s):\n", 
                   xml2::xml_text(content)))
-    # return(invisible(NULL))
+    return(invisible(NULL))
   }
   
   api_obj <- structure(
@@ -127,4 +116,58 @@ wfs_api <- function(base.url = NULL, queries) {
   api_obj$content <- content
   
   return(api_obj)
+}
+
+#' @title Gracefully fail internet resources
+#' 
+#' @description Function for gracefully failing internet resources
+#' 
+#' @author Kostas Vasilopoulos
+#' 
+#' @source Gracefully failing HTTP request code (slightly adapted by Pyry 
+#'    Kantanen) from RStudio community member kvasilopoulos. Many thanks!
+#'    
+#'    Source of the original RStudio community discussion:
+#'    \url{https://community.rstudio.com/t/internet-resources-should-fail-gracefully/49199}
+#' 
+#' @importFrom httr GET http_error message_for_status
+#' @importFrom curl has_internet
+#' 
+#' @keywords internal
+gracefully_fail <- function(remote.file, timeout.s = NULL) {
+  
+  if (is.null(timeout.s)) {timeout.s <- 10}
+  
+  try_GET <- function(x, ...) {
+    tryCatch(
+      httr::GET(url = x, timeout(timeout.s), ...),
+      error = function(e) conditionMessage(e),
+      warning = function(w) conditionMessage(w)
+    )
+  }
+  
+  is_response <- function(x) {
+    class(x) == "response"
+  }
+  
+  # First check internet connection
+  if (!curl::has_internet()) {
+    message("No internet connection")
+    return(invisible(NULL))
+  }
+  # Then try for timeout problems
+  resp <- try_GET(remote.file)
+  if (!is_response(resp)) {
+    message(resp)
+    return(invisible(NULL))
+  }
+  # Then stop if status > 400
+  if (httr::http_error(resp)) { 
+    httr::message_for_status(resp)
+    message("\n")
+    return(invisible(NULL))
+  }
+  
+  x <- "all ok"
+  invisible(x)
 }